(PUP-10606) Extract FFI from util/windows
Move FFI related functions, constants and structures out from
Puppet::Util::Windows, as loading Puppet::Util::Windows is extremely
expensive.
