+++ lib/puppet/ffi/windows.rb
@@ -1,0 +1,6 @@
+require 'puppet/ffi'
+module Puppet::FFI::Windows
+  require 'puppet/ffi/windows/structs'
+  require 'puppet/ffi/windows/functions'
+  require 'puppet/ffi/windows/constants'
+end
